requirements.txt
absl-py==2.1.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asciitree==0.3.3
asttokens==2.4.1
atomics==1.0.2
backcall==0.2.0
beautifulsoup4==4.12.3
blessed==1.20.0
bokeh==3.1.1
boto3==1.35.24
botocore==1.35.24
cachetools==5.3.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.4
colorlog==6.8.2
comm==0.2.2
commonmark==0.9.1
contextlib2==21.6.0
contourpy==1.1.1
cycler==0.12.1
dask==2023.5.0
debugpy==1.8.1
decorator
diffusers==0.26.3
docker-pycreds==0.4.0
docstring-parser==0.15
dt-apriltags==3.1.7
egl-probe==1.0.2
einops==0.8.0
evdev==1.7.0
executing==2.0.1
Farama-Notifications==0.0.4
fasteners==0.19
filelock==3.13.1
fonttools==4.49.0
fsspec==2024.2.0
fvcore==0.1.5.post20221221
gdown==5.1.0
gitdb==4.0.11
GitPython==3.1.42
glfw==1.12.0
google-auth==2.18.0
google-auth-oauthlib==0.4.6
gpustat==1.1.1
grpcio==1.54.0
gym==0.22.0
gym-notices==0.0.8
gymnasium==0.28.1
h5py==3.11.0
huggingface-hub==0.20.3
hydra-core==1.3.2
idna==3.6
imageio==2.31.3
imageio-ffmpeg==0.4.9
importlib-metadata==7.0.1
importlib-resources==6.1.1
iopath==0.1.10
ipdb==0.13.13
ipykernel==6.29.4
ipython==8.12.3
ipytree==0.2.2
ipywidgets==8.1.2
jax-jumpy==1.0.0
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
llvmlite==0.41.1
locket==1.0.0
manifold3d==2.4.5
Markdown==3.3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.6
mdurl==0.1.2
moviepy==1.0.3
mpmath==1.3.0
mujoco==2.3.3
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.1
ninja==1.11.1.1
numba==0.58.1
numcodecs==0.12.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.9.0.80
packaging==23.2
pandas==1.3.5
parso==0.8.3
partd==1.4.1
pathtools==0.1.2
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
platformdirs==4.2.0
portalocker==2.8.2
proglog==0.1.10
prompt-toolkit==3.0.43
protobuf==3.20.0
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.22
pygame==2.1.0
Pygments==2.17.2
pynput==1.7.6
PyOpenGL==3.1.6
pyparsing==3.1.1
pyrealsense2==2.54.2.5684
PySocks==1.7.1
python-dateutil==2.8.2
python-xlib==0.33
# pytorch3d==0.3.0
pytz==2023.3
PyYAML==6.0.1
pyzmq==25.1.2
rechunker==0.5.2
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rsa==4.7.2
s3transfer==0.10.2
safetensors==0.4.2
scikit-learn==1.3.2
scipy==1.10.1
sentry-sdk==1.40.5
setproctitle==1.3.3
Shimmy==1.3.0
shtab==1.6.4
six==1.16.0
smmap==5.0.1
soupsieve==2.5
stable-baselines3==2.0.0
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6.2.2
tensordict==0.4.0
termcolor==2.4.0
threadpoolctl==3.5.0
timm==0.9.16
tokenizers==0.15.2
tomli==2.0.1
toolz==0.12.1
# torch==2.3.0
# torchvision==0.18.0
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
transformers==4.38.1
trimesh==4.3.2
triton==2.3.0
typing_extensions==4.9.0
tyro==0.5.10
urllib3==1.26.18
wandb==0.16.3
wandb_osh==1.2.2
wcwidth==0.2.13
Werkzeug==2.2.3
widgetsnbextension==4.0.10
xyzservices==2024.9.0
yacs==0.1.8
zarr==2.16.1
zipp==3.17.0
